+++ Repositories/Implement/ProductRepository.cs
@@ -21,7 +21,7 @@ namespace Repositories.Implement
 
         public IEnumerable<Product> GetAllWithInclude(string field) => ProductDAO.Instance.GetAllWithInclude(field);
 
-        public Product? GetById(long id) => ProductDAO.Instance.GetById(id);
+        public Product? GetById(int id) => ProductDAO.Instance.GetById(id);
 
         public void Save() => ProductDAO.Instance.Save();
