+++ Repositories/Interface/IProductRepository.cs
@@ -12,7 +12,7 @@ namespace Repositories.Interface
         IEnumerable<Product> GetAll();
         IEnumerable<Product> GetAllWithInclude(string field);
         IEnumerable<Product> GetAllWith2Include(string field1, string field2);
-        Product? GetById(long id);
+        Product? GetById(int id);
         void Add(Product product);
         void Update(Product product);
         void Delete(Product product);
